created new service ProjectsService and removed methods in repos_projects that belogs ProjectsService
